Add Prometheus as datasource for Applications
This commit introduces the concept of datasources. Datasources are used
to retrieve metrics, logs or traces for an application and maybe later
also for other Kubernetes resources.

In the first step we are adding Prometheus as datasource for Application
metrics. For that the Application CRD was extended with a metrics
section, where a user can define variables and charts for the metrics
view of an application. The metrics section also contains a health field
with a single chart, which can be used to display the health of an
Application in the gallery overview.