Share datasource options between components
This changes the handling of the datasource options, so that they are
shared between the components for the Application details. This allows
us to use the same time range for metrics, logs and traces.

For that the datasource options are now handled in the parent component
of the different view, instead of the Metrics, Logs, Traces component.

This also allows us, to reflect the changes of the options in the
current URL via query parameters, so that a user can share his current
view with other users and they see the same time range.

Note: Maybe we can do sth. similar also for the users selected
variables, so that he also can share these values with other users.