Add page to directly query a Prometheus instance (#8)
Until now, it was only possible to use the Prometheus datasource, within
an Application CR. Now it is also possible to directly query a
configured Prometheus datasource within kobs. So that users havn't to
switch between kobs and the offical Prometheus UI to explore their data.

For this we added a new datasources page, where the user can select one
of his provided datasources, to use this datasource directly instead of
via the Application CR. The data for the page is provided by a new gRPC
method called GetDatasources, which returns all configured datasources.

To make this new feature as easy as possible, we also added a new
actions dropdown to each chart, where the user can select his query and
is redirected to the corresponding datasource page, so he can dig deeper
into his data.