Fix time parsing for the datasource options
The user provided time string wasn't parsed correctly. We used the
provided start time also for the end time by mistak. Also the provided
string was always parsed as UTC date instead of the users current
timezone. Both issues are fixed now.

We also added a new input field for the Prometheus resolution. This
option was already present, but the input field was missing. So that a
user wasn't able to change the resolution. The resolution is displayed
in a fourth column in the options modal. This column should also be used
for other datasources (e.g. max results for Jaeger).