Add Elasticsearch as datasource
We are adding Elasticsearch as second datasource besides Prometheus.
Elasticsearch can be used to query the logs of our applications. For
that we extend the Applications CRD with a logs field, which can be used
to define Elasticsearch queries to retrieve the logs of an application.

In the kobs UI we added a new tab "logs" for the applications view,
where a user can select the defined queries for an application. When a
query is selected, the query is run against Elasticsearch and the logs
are shown in a table view. Next to the query we also run an aggragation,
so show the distribution of the logs in the selected time range in a bar
chart.

Next to the support of Elasticsearch in the applications view, we also
added an interface to query the configured Elasticsearch datasource
directly.