Refactor cluster and application handling (#11)
Refactor the cluster and application handling, for a better development
experience. The Go code for the clusters and applications lives now in a
plugins folder, to prepare the gRPC server for our new plugins concept.

All React components for the clusters and applications were reworked,
because the old components, were following an ugly naming schme. Also
some of the components were to complex from the logic they had to do.
The new components should improve the experience with the frontend code
for kobs.