Add plugin system
This commit introduces a new plugin system for kobs. The new plugin
system can be used to extend the functionality of kobs. Each plugin is
described by his own ".proto" file. The plugin can then be registered
in the "pkg/api/plugins/plugins/plugins.go" file. The frontend
components must be registered in the "app/src/utils/plugins.tsx" file.
Each plugin must define two React components: One component is
responsible for directly using the plugin. THe other component allows
the usage of the plugin within an application.

We also readding Prometheus and Elasticsearch as plugins (previously
known as datasources). The Prometheus plugin allows a user to define
variables and charts within the Application CR, The metrics are then
shown as charts under the corresponding plugin tab. Next to this the
plugin also allows a user to directly query a configured Prometheus
instance. The Elasticsearch plugin allows a user to define queries
within the Application CR. The retrieved logs are then shown in the
corresponding plugin tab. The Elasticsearch plugin also allows to
directly query an Elasticsearch instance.