Add Jaeger Plugin (#13)
Add a new plugin, which allows a user to retrieve and view traces from
Jaeger. For the Jaeger plugin we have added a new "jaeger" section to
the plugins section in the Application CRD. This allows a user to
specify the "service", "operation" and "tags" within his CR. The traces
can then be viewed in the corresponding plugin tab.

The Jaeger plugin can also be used as standalone plugin, without an
Application CR. The Jaeger plugin is the first plugin, which supports
multiple pages for the standalone mode. The first one allows the user to
retrieve traces, by specifying a "service", "operation", "tags", etc.
simular to the Jaeger Query dashboard. The second page can be used to
view a single trace and to compare this trace with another trace.