Fix loading of Jaeger services
When a user opened the Jaeger plugin page, where the "serivce" query
parameter was already present in the URL, the list of services wasn't
loaded from our gRPC API.

To fix this issue, we added a new function "GetServices" and split the
gRPC call for services and operations, so that they are now loaded
independently.

We also added a new "All Operations" options to the select box for the
operations, so that it is possible to unselect an already selected
operation via the new "All Operations" option.