Use location to load applications
Similar to the plugins, we do not load the applications directly anymore
after the user clicked the search button. Instead we change the
location.search parameter for the applications page. For each change to
the location the "fetchApplications" function is triggered to load the
applications. This allows a user to share his current selection of
clusters and namespaces with other users.