Use multiple colors in the Jaeger plugin (#19)
The Jaeger plugin uses multiple colors now. Right after a trace was
returned from the gRPC service, we loop through all traces and their
processes and asign each process a color. This color is then used in the
list of spans for the service and in the chart of spans and spans view.

This also fixes a bug, where the specified traces where not used in the
API. The tags specified by the user will now be used for the Jaeger API
calls. For that we transform the string into a json object. This object
is then encoded and the encoded string is sent to our gRPC server.

Besides the mentioned changes from above, we also added two function, to
check if a trace/span contains a tag "error=true". If this is the case,
we display a error icon next to the trace/span id in the title.