Rework usage of icons, links handling and drawer
Until now we had used the icons by importing them directly from their
corresponding file. Now we are using the "@patternfly/react-icons"
package to import the icons.

The links handling wasn't unified accross the frontend. Now we are using
a new approach by wrapping components in a new "LinksWrapper" component
so that each link is opened within kobs by default. This allows the user
to decide if he wants to open a link in a new tab or in the current one.

The drawer layour was completly changed. We had a helper variable
"isInDrawer" to apply a specific layout, when a component was displayed
within the Apllication drawer. This variable was removed. Instead we are
always using the same layout. We also removed the "isFlat" property from
most of the "Card" components to have a consistent style accross all
components. Last but not least we wrapped most of the components in a
"Card" component to be consistent with other components, which were
already displayed in cards.