Add preview for Applications via plugins (#21)
This commit adds a new "plugin" field to the details section of the
Application CR. This allows a user to specify a plugin, which is then
used as preview for the Application in the Application overview page.

At the moment a user can specify a Prometheus or Elasticsearch query for
the preview. When this is specified the corresponding Prometheus chart
with the metrics for the query, or the number of hits and the buckets
chart will be displayed in the Application card. If the user doesn't
specify a plugin, we will display the provided description of the
Application.

This commit also fixes a small bug, where the list of resources wasn't
updated correctly, because we missed to compare the properties in the
"React.memo" function.