Add Helm chart (#22)
Add a Helm chart to deploy kobs into a Kubernetes cluster via Helm. This
is a first very basic version of the Helm chart and should be extended
later.

The Helm chart can be found in the "deploy/helm/kobs" folder. Every time a
file in this folder is changed, we trigger the corresponding GitHub
action to run "helm lint", "helm template" and "helm install" against a
kind cluster.

If the Helm chart is changed on the main branch we automatically trigger
a release and publish a new version of the Helm chart in the
"kobsio/helm-repository" GitHub repository. The Helm chart can then be
used via https://kobsio.github.io/helm-repository/.