Add Kustomize files and demo
This commit adds the required files to deploy kobs using Kustomize. The
files are generated using "helm template kobs deploy/helm/kobs" and
manually adjusted afterwards. To deploy kobs using Kustomize the
following command can be used: "kustomize build deploy/kustomize |
kubectl apply -f -".

We are also adding a demo application, to demonstrate the usage of kobs.
The demo deploys the Istio Operator and Istio, the Bookinfo example from
Istio and the following observability tools: Prometheus, Elasticsearch
and Jaeger. To deploy kobs we are using the Kustomize files, but we
overwrite the default configuration, to enable the Prometheus,
Elasticsearch and Jaeger plugin.