Add support for all Kubernetes resources (#26)
We are supporting all Kubernetes resources. This means a user can also
view ConfigMaps, Secrets, ClusterRoles, ClusterRoleBindings, etc. within
kobs.

We also adjusted the icons for applications and plugins on the homepage.
We fixed the header high in the compare traces component of the Jaeger
plugin. Last but not least we fixed the handling of external links in
the LinkWrapper component.