Add dependencies section and topology graph (#29)
The Application CRD contains a new field "dependencies" now. This field
can be used to specify other applications as dependency for an
application. This dependecies are then used during the startup of kobs
to generate a list of nodes (clusters, namespaces, applications) and a
list of edges (dependencies between applications). The list of nodes and
edges is then used to create a topology graph using Cytoscape.js.

The user can switch the views in the applications page between the
gallery mode and the new topology view. In the same as it is in the
gallery view, the user can filter the applications by cluster and
namespace. A user can also select an application (node) in the topology
graph to show the drawer with the application details.

We also adjusted the demo for the new topology graph feature. For that
we have added more Application CRs and added a dependency section to all
CRs.

To add this new feature we also had to move the logic for fetching the
applications to the gallery component. We also renamed the component to
have a consistent style for the component nameing.