Add plugin support for Kubernetes resources
It is now possible to use plugins within Kubernetes resources. For that
a user can specify the "kobs.io/plugins" annotation on each resource and
kobs will add the specified plugins, when a user views the resource. The
plugin annotation must be a list of plugins, which have the same
specification as the plugins section in the Applications.

It is also possible to specify a list of Applications within a resource.
For that the "kobs.io/applications" annotation can be used. The
annotation requires a list of object, where each object must specify the
"name" of the application and an optional "namespace", when the
Application isn't in the same namespace as the resource.