Add support for container logs via Kubernetes API
If a user selects a Pod in the resource view, a new tab "Logs" is now
available. Within this tab the user can retrieve the logs for a selected
container of the Pod. Besides the container a user can also select the
time in seconds since when the logs should be retrieved, a regular
expression to filter the returned logs and an option to get the logs of
an previous instance of the Pod.

Notes:

- All coponents are wrapped in a Card component now (also the Yaml
  editor and the events table), so that the style is consistent with the
  plugins.
- We have added the "apiVersion" and "kind" fields to the Pod
  representation, so that we can check if the selected resource is a
  Pod. Maybe we should also do this for all the other Kubernetes
  resources.