Fix topology chart for invalid dependency data
When an application has an non existing application defined as
dependency, the topology component in the React UI crashed. To avoid
this we loop through all edges, when we have collected all data for the
topology chart and remove all edges, which do not contain a valid source
or target.