Add a new Custom Resource Definition for Teams
We are intoducing a new Custom Resource Definition for Teams. Teams can
be used to define the ownership of an Application or for other
Kubernetes resources via the "kobs.io/teams" annotation. Currently the
teams can only contain a description, logo and a list of links.

The Application CR also contains a new field "teams", where a user can
define a list of team names. The name of each team is then rendered next
to the defined links of the Application.

We also introduced a new ExternalLink component, which is used to render
the links defined in an Team or Application CR. We introduced this new
component to add the ExternalLink icon to each link or a special brand
icon, when the link contains the brand (e.g. github.com).

Finally we are using the "apiextensions.k8s.io/v1" API for the
Application and Team CRD now, because the old API will be removed in the
next Kubernetes version (1.22) and we do not want to care to much about
backwards compatibility.