Allow users to overwrite the plugin name (#35)
Until now we always used the plugin name in the tabs list for
Applications and Kubernetes resources. Now it is possible to overwrite
this name with the new "displayName" field.

This could be useful in cases where one plugin is used multiple times
within an Applications or resource. For example the Prometheus plugin
can be used multiple time: In the first tab it is used to display the
application metrics and in a second tab it is used to display the Istio
metrics for this application.