Add Opsgenie plugin
This commit adds a new plugin for Opsgenie. The plugins allows a user to
view the alerts from a configured Opsgenie instance within kobs.

The plugin can also be used within an Application or Kubernetes
resource. For that a user can define a list of queries, to retrieve
alerts from the Opsgenie API.