Prometheus metric name suggestion (#40)
The TextArea component in the Prometheus plugin now supports
suggestions for the entered PromQL query. To get the suggestions, the
entered query is send to the gRPC API and compared with all the
available time series. The time series, which match the entered query
are returned to the frontend and shown as a suggestion list below the
TextArea component. A user can then select one of the items via mouse
or the arrow up/down and enter key.

Co-authored-by: Falk Scheerschmidt <falk@staffbase.com>
Co-authored-by: Matthias Munder <matthias.munder@staffbase.com>
Co-authored-by: ricoberger <mail@ricoberger.de>