Add overview and Pods tab for resource details
This commit adds a new overview tab for all resources in the details
drawer. The overview tab shows the metadata and other important fields
of a resource in a clear way. It should help the user to get a faster
overview of the current status of a resource.

For other resources, like Deployments, StatefulSets, etc. we also add a
Pods tab. The Pods tab shows the pods for a resource, which are matching
the specified label selector.

Last but not least, we are adjusting the query parameters in the
resource view, when the user selects a cluster, namespace or resource,
so that a user can share the url with other users. This also allows us
to link to specific resource by providing a cluster, namespace, resource
and selector in the URL.