Add VirtualService for Istio to the Helm chart
If kobs is deployed in a cluster, where Istio is running, the Envoy
sidecar is not needed, because the gRPC web traffic can also be handled
by the Istio sidecar.

To improve the user experience with such a setup, it is now possible to
create a VirtualService to expose the kobs dashboard. For that we also
had to rename the port for the gRPC traffic to "grpc-web", so that Istio
can detect the used protocol.