Add value mappings for sparkline charts (#45)
Sparkline charts in the Prometheus plugin are supporting value mappings
now. This allows a user to specify a map to overwrite the displayed
value.

We also fixed a bug for NaN values in the charts. The used charting
library doesn't support NaN values, so we have to convert them to "null"
before the chart is rendered.