Add new chart type table for Prometheus plugin (#49)
The new "table" chart type for the Prometheus plugin allows a user to
render the result of multiple Prometheus queries in a table. To use the
"table" chart type a user must also specify a list of "columns" in the
plugin soecification.

If the table is rendered with multiple queries, the queries are joined
by the specified "label".