Fix determination of the root span
This fixes the determination of the root span in the Jaeger plugin. In
some cases it can happen that the first span isn't the root span, so
that we have to loop through all span. Then we return the first one,
which doesn't have any references with an "CHILD_OF" entry.