Add command-line flag to forbid resource access (#51)
This commit adds a new command-line flag
"--clusters.forbidden-resources", which can be used to forbid access to
a list of resources. This allow a user to still use RBAC with wildcards
for the permissions of kobs, but do not allow a user to access the
specified resources.

We also adjusted the corresponding React component, so that the error is
now shown to the user. This also  fixes a problem, where the app
crashed, whenn the empty state was clicked in the resources list.