Fix fields handling in Elasticsearch plugin
Fix a bug, where the selected fields were removed for a new search. Now
we are keeping the fields across multiple searches. For that we are
parsing the location and reuse all existing fields.

Fix a bug, where the fields list wasn't created, when the search
returned less then 10 documents.