Add tags as filter in the Jaeger plugin (#55)
It is now possible to directly add a tag from a span as filter in the
Jaeger plugin. To do so, a user has to click on the tag and then he must
click on the "Filter" link. In the current logic all existing tags are
replaced by the selected tag.

We also adjusted the compare page, by directly using react-dropzone
instead of the FileUpload component from Patternfly. This allows us to
have a cleaner layout on the page.