Add action for Elasticsearch plugin
The new actions for the Elasticsearch plugin can be used to
include/exclude a value from the search and to toggle a field in the
table view. These actions are only available in the page view of the
Elasticsearch plugin and not within the Applications view.

We also fixed a small bug, where double quotes were not escaped
correctly in the request against the Elasticsearch API.