Add support for templates (#59)
It is now possible to define templates for plugins, so that plugins can
be reused accross Applications, Teams and Kubernetes resources.
Templates are defined via the Template CRD. The name of a template must
be unique accross clusters and namespaces. Every template can contain
variables, which then must be provided within the usage in an
Application or resource. Variables in templates must have the following
format: "<< VARIABLE_NAME >>". When the template is used within a
resource the value of a variable can be set via JSONPath.