Improve caching logic
This commits changes the behaviour of the current caching logic for
teams and the topology graph. We are using a similar logic as it was
intorduced for the plugin templates. This means that the teams and
topology data isn't generated in an additional goroutine, instead we get
the data on the first request and then we cache it for the defined cache
duration.