Add Kiali Plugin (#63)
The new Kiali plugin can be used to visualize an Istio service mesh
within kobs. For that a user can select a list of namespaces on the
plugin page. We then request the graph data from the Kiali API and
render the Graph using cytoscape.

Currently the plugin only supports the graph feature of Kiali, we plan
to extend the plugin later. It is planned to also retrieve the health
and metrics for each service.