Add edge metrics for Kiali plugin
When the user selects an edge from the Kiali graph, we now show a drawer
with additional details for the selected edge. The details view contains
the same tabs as the Kiali UI, to show the traffic details and the
traffic details by flags and hosts.

For this we have introduced a new gRPC method for the Kiali plugin
called "GetMetrics". The method can be used for edges and late for nodes
to get the Prometheus metrics for a users selection.