Readd teams and applications
In this commit we are readding the Custom Resource Definitions for Teams
and Applications. The new implementations doesn't rely on protobuf
definitions anymore. Instead we are defining all the structs and fields
directly in the api definition. The CRDs are very similar to the old
ones, but containing some breaking changes, for that we have decided to
bump the CRD versions to v1beta1.

The teams and applications are now also implemented via a corresponding
teams and applications plugin. The plugin supports the same functions as
the original implementation of the teams and applications components,
except that we do not support plugins directly anymore. Instead of
directly using plugins within the teams and applications, we will use
the upcoming Dashboards Custom Resource Definition, which can be used by
the teams and applications.