Add Custom Resource Definition for Dashboards
This commit adds a new Custom Resource Definition, which allows users to
create dashboards. Dashboards should replace the existing CRD for
templates and the reference of plugins within Teams, Applications and
Kubernetes resources.

Each dashboard contains a number of rows and each row contains a number
of panels. The panels can be used to add a plugin to a dashboard. For
this the user must provide the name of the plugin and some optional
options. The options are passed to the Panel component of the plugin and
therefore must be valided in the Panel component of the plugin.

We also pass the cluster, namespace and name of the team / application
is used as reference as property. This allows a user to omit the cluster
and namespace in some plugin options.