Add support for placeholder in the Dashboards CRD (#75)
The Dashboards CRD and the references in the Teams and Applications CRD
are supporting placeholders now. Placeholders can be to create more
general usable dashboards and let the user decide which value should be
used for the placeholder.

For that we are adding a placeholders array to the Dashboards CRD, which
is mainly used to show all the available placeholders later in the list
of dashboards. The value for a placeholder is set in the placeholders
map of the Teams/Applications CRD. These values are passed to the Go API
and when we retrieved a dashboard for a given reference we are using the
template package to replace the placeholders in the dashboard with the
correct value. The placeholders can be specified via the following
format: "{% .placeholder %}".