Add support for variables and times in dashboards (#76)
It is now possible to define variables for a dashboard. In contrast to
placeholders variables can be used to change the content of the
dashboard panels during runtime, while placeholders must be defined
beforehand. Currently we are supporting three variables options, one to
select a cluster, one to select a plugin and one, where a user can
provide a static list of possible values. Variables can be used with the
following format: "{% .variable %}". For that we had to change the
delimiter for placeholders to "{{ .placeholder }}". In the future we can
extend the variables to make usage of other plugins (e.g. Prometheus).

Next to the variables it is now also possible to select a time range for
a dashboard. This time range is then passed to all the panels with there
plugins, so that for example the Prometheus plugin can show metrics for
the selected time range.

We also changed the PluginCard component so that it is now possible to
have a transparent background for the panels. This is useful for plugins
which are already using cards to display there data, like teams,
applications, dashboards and resources.