Re-add Prometheus Plugin
Re-add the Prometheus plugin as it already exists. We adjusted the
existing Prometheus plugin to fit in the new plugin system for kobs. For
that the code for the plugin lives now in the plugins folder.

The Prometheus plugin supports the same features as the former
implementation, but we are now using nivo to render the charts instead
of the Patternfly charting library. We decided to change the library,
because nivo can render canvas charts, which comes with a better
performance for the dashboards.

We also made some adjustments to the dashboards handling, so that we are
now only calculating the dimensions for the dashboards once instead of
for every dashboard. We also adjusted the tab handling, so that the
currently selected dashboard is reflected in the url. We plan to extend
this later, so that also the selected time range and variables are
reflected in the url.