Re-add Elasticsearch Plugin (#78)
We are re-adding the Elasticsearch plugin, which now uses the new plugin
functions of kobs. The plugin page supports the same features as the
original Elasticsearch plugin. The plugin can now be used in a
dashboard. For that the user can define a query via the options for the
Elasticsearch panel. A user can also add multiple fields which will then
be shown in a table format. Last but not least the user can also set a
boolean value to show the chart from the Elasticsearch page also in a
dashboard panel.