Re-add Jaeger Plugin
Re-adding the Jaeger plugin to view traces from a configured Jaeger
instance. The plugin supports the same features as before, which means
you can use the Jaeger page to get traces by a service, operation and
tags. You can also compare two traces.

We added a new Jaeger panel, which allows a user to include traces into
a dashboard. For that a user can specify the same properties in the
Dashboard CR as he can do on the Jaeger page.