Add Preview Components and Dashboard Page
It is now possible to define a preview for Applications. For that a user
can set the plugin and the plugin options which should be used for the
preview. Currently only the Prometheus and Elasticsearch Plugin supports
the preview feature. If the user haven't specified a preview component
we show the description for the application. If the description wasn't
provided the cluster and namespace will be shown.

It is now possible to view a single dashboard via the dashboards page.
When the user selects a dashboard, he must provide the default cluster
and namespace for the dashboard and a value for all the placeholders
which are defined within the dashboard. These values are then passed as
query paramters to the dashboards view.