Rework Documentation, Helm Chart and Kustomize
We had to adjust the documentation for the new version of the Teams and
Applications CRD. We also had to document the new Dashboards CRD and how
plugins can be added to a dashboard.

Since we removed protobuf as dependency we do not need Envoy anymore.
For that we removed Envoy from the Helm chart and Kustomize files.

We also did some other small adjustments, like updating the dependencies
and adding actions for the applications, resources and Jaeger panel.