Add markdown plugin
Add a markdown plugin, which can be used to render static text within a
dashboard panel. The plugin only implements the panel component, a
corresponding page component isn't available.

For the missing page component we improved the message which is shown
when such a component is not implemented in the core package.

Last but not least we fixed the GitHub Action for Helm, so that the Helm
chart is only published on commits to the main branch but not for
created tags.