Improve error handling (#82)
This commit improves the error handling in our Go API. To return an
error we can now directly call errresponse.Render, which sets the passed
in status and returns the message and error as an JSON object to the
user. This error will now also be logged in our custom request log
handler.