Add special annotations for Kubernetes resources (#83)
It is now possible to use teams, applications and dashboards within all
Kubernetes resources via some special annotations. The "kobs.io/teams"
annotations can be used to specify a list of teams for a resource. The
annotation requires a list of teams, with the cluster, namespace and
name of the team. If the cluster/namespace isn't set the
cluster/namespace of the resource will be used. The
"kobs.io/applications" annotation can be used in the same way to specify
a list of applications for a resource. Last but not least the
"kobs.io/dashboards" annotation can be used to set a list of dashboards
which should be shown for the resource. If the dashboards uses
placeholders, the values for these must also provided. To set the value
for a placeholder it is also possible to use a JSONPath to set the value
to the value of a property in the Kubernetes manifest file for the
resource.