Improve overview page for Pods
The overview page for Pods now, displays all Containers in an expandable
table. When a container is selected all the details for the selected
container are displayed below this table row in a description list like
the Pod details.

We also try to get the CPU and memory usage for all containers. If the
API call returns the resources, we display them in the containers table.