Improve overview page for Nodes
This commit adds a separate overview page for nodes, where we show the
requests, limits, usage and capacity for the cpu and memory of the node.
We also show the capacity and usage for pods. This data is shwon in
three bar charts, where each bar indicates one value.

For node we are also showing the Pods tab in the details, which contains
all the Pods running on this node.