Rework Kiali plugin (#87)
This commit adds the reworked Kiali plugin. The Kiali plugin can be used
to show the topology graph and metrics from a configured Kiali instance.
For that a user can select a list of namespaces, for which the topology
graph is then generated.

In the reworked version of the plugin it is now also possible to show
the topology graph within a dashboard.

Currently only the topology graph is supported. We plan to add charts
for the metrics from Kiali in the future.