Improve handling of actions for resources
We are now refreshing the list of resources, when the user triggers an
action for the currently selcted resource. Besides that we also reworked
the shown alerts, so that it is now possible to show multiple alerts at
one time. We also show a success message, so that the user knows that
his action succeeded.