Rework Opsgenie Plugin (#89)
This commit adds the reworked version of the Opsgenie plugin, which is
known from the version 0.3.0 of kobs. With the reworked Opsgenie plugin
it is possible to get a list of alerts and incidents from Opsgenie and
embed them in a dashboard.

A user can provide a query to get a list of alerts and incidents from
Opsgenie. kobs automatically adds a time restriction for the list of
alerts based on the selected time range of the dashboard or the Opsgenie
page.