Add force delete option for Kubernetes resources
It is now possible to force delete Kubernetes resources. To use this
options the user has to check the corresponding checkbox in the delete
modal. We then set the force parameter to true and creating the
corresponding request body in the backend to let the Kubernetes API now
that a resource should be force deleted.

This commit also fixes a bug in the Opsgenie plugin, where it was not
possible to query for all alerts/incidents. This is now possible,
because we only add the default "status: open" query, when the query
parameter is null.