Preparation to build a own kobs version
This is the preparation for the kobsio/app respository, which allows
users to build their own version of kobs. This is useful, so that users
can decide which plugins they want to use within kobs and that users can
build their own private plugins.

To support this function we moved the config and plugins package of kobs
from the pkg folder to the cmd folder. The cmd folder is then reused in
the kobsio/app repository and allows a user to add his own plugins to
the plugins.go file. To add the frontend code of a plugin the user can
modify the index.tsx file from the app folder, where we just added a new
css file with a fix for the Patternfly Design System.