Add column to show Kubernetes resource status (#93)
For some of the Kubernetes resources we added a new column to the end of
the table to show the current status of the resource. This allows a user
to see if there is a problem with a resource within the table view. So
that the user do not have to open the details view for the selected
Kubernetes resource.