Get Kubernetes resources for all namespaces
It is now possible to get Kubernetes resources for all namespaces, by
not selecting a namespaces from the select box on the resources page.

We also fixed a bug, where it was not possible to get custom resources
for a CRD, because the request path for all CRDs was wrong.

In the Jaeger plugin it is now possible to copy the value of a tag by
clicking on it.