Add RSS Plugin
The new RSS plugin can be used to get the latest status updates for a
list of RSS feeds. The plugin can only be used within a dashboard. A
user must provide a list of urls and an optional sortBy paramter. The
urls are then send to the Go API and we try to retrieve all items from
the feed using the gofeed package. Before we return the items to the
React UI, we are transforming the structure of the feeds and sort them
by the provided field. The returned feed is shown as list in the React
UI. In the details view we are using the html-react-parser to show the
description of an item from the feed.