Add support for Kiali metrics (#97)
The Kiali plugin supports all the metrics from Kiali now. This means
when a user selects an edge or node in the Kiali topology graph he also
sees the metrics for TCP, HTTP and gRPC traffic. To get these metrics we
are using the 'metrics' endpoint of the configured Kiali instance.

Instead of the custom duration dropdown we are now using the Option
component from the core package to select the duration (time range) for
which the graph should be rendered. It is important to now that the time
range only works for the metrics, for the values in the topology graph
only the difference between the start and end time is used.