Add terminal support for Kubernetes Pods (#98)
It is now possible to create a terminal for a Kubernetes Pods. The
terminal can be created via the actions in the Pod details drawer. We
are using xterm.js for the terminal, all the input / output is send over
a WebSocket connection.

By default we are using the current domain for the WebSocket connection.
This can be overwritten in the configuration of the resources plugin for
kobs.