Fix WebSocket connection for Kubernetes terminal
This commit fixes the WebSocket connection for the Kubernetes terminal
in environments where a idle timeout for the connections is defined. To
avoid that the connection is closed when the WebSocket connection is
longer open then the specified idle timeout, we added a ping / pong
implementation. These should avoid that the connection is closed.