Add support for Ephemeral Containers (#100)
It is now possible to view Ephemeral Containers in the overview page of
a pod and to create Ephemeral Containers via the debug actions. To make
it easier for users to create such containers it is possible to specify
a list of templates for the Ephemeral Containers in the configuration
file via the 'ephemeralContainers' key in the resources plugin section.

To make use of an created Ephemeral Container we also added them to the
terminal modal. This allows our users to exec into the created
containers.

To support the parsing of these templates in from the configuration
file, we had to switch our used yaml package from gopkg.in/yaml.v2 to
sigs.k8s.io/yaml, which also supports parsing of yaml files via the
specified json tags.